Modified MorePlaces div and dom element labels.
